requirements.txt
numpy==1.19.5
pytest==7.0.1
tensorflow-gpu==2.6.0
keras==2.6.0
scikit-learn==0.24.2
pandas==1.1.5
matplotlib==3.3.4
shap==0.39.0